refactor: move shared auth helpers and constants to /shared
