+++ src/features/auth/auth.route.js
@@ -60,11 +60,6 @@ router.put(
   authController.changePassword
 );
 
-// router.post(
-//   "/onboarding",
-//   validate(authValidation.onboarding),
-//   authController.onboard
-// );
 router.post("/logout", authController.logout);
 
 module.exports = router;
